requirements.txt
docopt
requests
docker-py
